Fix for Extended APDUs
Not all Java Cards that support extended APDUs have APDU buffer
lengths long enough to include the attestation certificate. This
causes registration to fail when extended APDUs are used, as is the
case with iOS.

The fix is to build the extended responses incrementally instead of
staging them all at once in the APDU buffer.